40.1.DTO Input "Initializer"
Setting OBJECT_TO_POPULATE to the Input DTO
Inside the denormalizer, let's start with something simple: $dto = new CheeseListingInput() and $dto->title = some hardcoded title:bject, it's ignored and a new CheeseListingInput is created
But... we could change that key.
Set this onto the context: $context[AbstractItemNormalizer::OBJECT_TO_POPULATE] equals $dto
We're not done yet... but if we passed this  into the denormalizer system, then it should deserialize the JSON onto our new object. And, whatever we return from this method will eventually be passed as the  to our data transformer.
So next, let's finish this: by calling the denormalizer system to update the CheeseListingInput, returning it from here, proving that it's passed to the data transformer, then finally pre-filling it with database data. If you can't see how all the pieces connect yet, you will soon.