41.1.DTO Input Initializer Logic
The end goal is that this $input argument will be the CheeseListingInput object that we create in our denormalizer... which eventually will be pre-filled with data from the database:

 42 lines  src/DataTransformer/CheeseListingInputDataTransformer.php
class CheeseListingInputDataTransformer implements DataTransformerInterface
{
    public function transform($input, string $to, array $context = [])
    {
    }
}
Let's dump($input) so that, once this ready, we can see if we've accomplished that

Calling the Core Denormalizer System
As I mentioned earlier, as soon as we created this class and filled in supportsDenormalization():

 30 lines  src/Serializer/Normalizer/CheeseListingInputDenormalizer.php
class CheeseListingInputDenormalizer implements DenormalizerInterface, CacheableSupportsMethodInterface
{
    public function supportsDenormalization($data, string $type, string $format = null)
    {
        return $type === CheeseListingInput::class;
    }
}
We became 100% responsible for denormalizing CheeseListingInput objects... which means we actually do need to somehow take the array of data from the JSON, put it onto the CheeseListingInput and return it!
But... pfff. We don't really want to do all that hard work... because that's what the normal denormalizer system does. Nope, let's just call the normal system, but pass in our modified $context.
We actually had this same situation in an earlier tutorial with UserNormalizer and we handled it with a - kind of - elaborate solution that allowed us to inject the entire normalizer system, call normalize() again and use a flag on the  to avoid recursion:

 67 lines  src/Serializer/Normalizer/UserNormalizer.php
class UserNormalizer implements ContextAwareNormalizerInterface, CacheableSupportsMethodInterface, NormalizerAwareInterface
{
    use NormalizerAwareTrait;
    public function normalize($object, $format = null, array $context = array())
    {
        [self::ALREADY_CALLED] = true;
        return $this->normalizer->normalize($object, $format, $context);
    }
    public function supportsNormalization($data, $format = null, array $context = [])
    {
        // avoid recursion: only call once per object
        if (isset($context[self::ALREADY_CALLED])) {
            return false;
        }
        return $data instanceof User;
    }
}
The most correct solution in CheeseListingInputDenormalizer would be to do that same thing. But... I'm going to cheat. Denormalization is a bit simpler than normalization... and we can get away with injecting the specific, one denormalizer that I know we need instead of injecting the entire denormalizer system and trying to avoid recursion.
The denormalizer we need is called ObjectNormalizer and it's autowireable. On top, create public function __construct() with ObjectNormalizer - make sure you get the one from Symfony - $objectNormalizer
I'll hit Alt+Enter and go to "Initialize properties" to create that property and set it
Now, down in our code, return $this->objectNormalizer->denormalize() and pass it the $data - because we still want to denormalize the same array of data - $type \format, but now pass our shiny new $context so that when it denormalizes, it will update our $dto instead of creating a new one.
Ok, let's try it! Hit "Execute" again. We still get a 500 error... which makes sense because we haven't finished our job yet. What I want to see is if the dump worked: I want to see if the CheeseListingInput object we created in the denormalizer is now being passed into the transform() method.
Go over to the other profiler tab and hit "Latest". And... there it is! Oh, wait. Ryan, you dummy! I should have taken off the title field from the JSON so that we can see if the title that we're setting in the denormalizer is what we see in the dump(). Try it again... 500 error... hit Latest and... yes! This is what we were looking for! It proves that we are now in control of creating the CheeseListingInput object. We create the object, then, when the deserializer does its job, it either leaves the property alone if we didn't send that field or it overrides it if we did send that field.
Now we are dangerous. Inside CheeseListingInputDataTransformer, once we've finished pre-populating the $input object from the database in the denormalizer, we will be able to safely transfer every property onto CheeseListing because if a field was not sent in the JSON, it will match what's already in the database.

OBJECT_TO_POPULATE, CheeseListing & CheeseListingInput
By the way, if you're wondering why we can override OBJECT_TO_POPULATE to be a CheeseListingInput inside the denormalizer:

 40 lines  src/Serializer/Normalizer/CheeseListingInputDenormalizer.php
class CheeseListingInputDenormalizer implements DenormalizerInterface, CacheableSupportsMethodInterface
{
    public function denormalize($data, string $type, string $format = null, array $context = [])
    {
        $dto = new CheeseListingInput();
        $context[AbstractNormalizer::OBJECT_TO_POPULATE] = $dto;
    }
}
But then OBJECT_TO_POPULATE is still apparently a CheeseListing object inside of our data transformer:

 42 lines  src/DataTransformer/CheeseListingInputDataTransformer.php
class CheeseListingInputDataTransformer implements DataTransformerInterface
{
    public function transform($input, string $to, array $context = [])
    {
        if (isset($context[AbstractItemNormalizer::OBJECT_TO_POPULATE])) {
            $cheeseListing = $context[AbstractItemNormalizer::OBJECT_TO_POPULATE];
        }
    }
}
That's... a good thing to wonder!
In reality, denormalizing into the input object and calling the data transformer are two completely separate processes. The two $context arrays - while nearly identical - are two separate arrays in memory. So even though we modify the $context inside the denormalizer, that does not modify the $context that's passed to the data transformer.