41.1.DTO Input Initializer Logic
Pre-Populating from the Database
Anyways, let's finish the denormalizer by populating the CheeseListingInput with the data from the database. And... since this is pretty boring, at the bottom, I'll paste a new private function. You can copy this from the code block on this page.
Re-type the end of CheeseListing to get that use statement.
This checks the object_to_populate key - I should probably use the constant - to see if there is an existing entity, which would happen for an update operation. If there is no entity, which means this is a create operation, it returns an empty DTO. But if there is a CheeseListing, it uses it to pre-populate all the fields on the DTO.
Back up in denormalize(), delete the new DTO stuff and say $context[AbstractNormalizer::OBJECT_TO_POPULATE] equals $this->createDto() and pass $context
Let's try it! Back at the browser, go to the docs tab... and hit Execute. Let's see... yes! It works! Price 5000!
We're still missing a few pieces related to validation, but we'll talk about those soon.
Before we do, we currently have code for converting to and from entity and DTO objects... all over the place. Next, it's cleanup time!