44.1.Input DTO Validation
One nice thing about input DTOs is that after our data transformer is called and we return the final CheeseListing, that object is validated like normal. We saw this: we submitted empty JSON to create a new CheeseListing and got back errors like "the title should not be blank".
These are coming from the @Assert rules on CheeseListing. CheeseListing is still validated.
But... this isn't the only way that validation can work. One complaint that you'll sometimes here about Symfony's validator is that, for it to work, you need to allow your entity to get into an invalid state. Basically, even though $title should not be blank
We need to first allow a blank or null value to be set onto the property so that it can then be validated.
This was at the root of a problem we had a minute ago. In CheeseListingInput, we had to add some type-casting here to help us set invalid data onto CheeseListing without causing a PHP error... so that it could then be validated:

 83 lines  src/Dto/CheeseListingInput.php
class CheeseListingInput
{
    public function createOrUpdateEntity(?CheeseListing $cheeseListing): CheeseListing
    {
        if (! $cheeseListing) {
            $cheeseListing = new CheeseListing((string) $this->title);
        }
        $cheeseListing->setDescription((string) $this->description);
        $cheeseListing->setPrice((int) $this->price);
    }
}
Moving the Constraints to the Input
Another option is to move the validation from the entity into the input class. If we did that, then when we set the data onto this CheeseListing object, we would know that the data is - in fact - valid.
So let's try this. Undo the typecasting in CheeseListingInput because once we're done, we will know that the data is valid and this won't be necessary
Next in CheeseListing, I'm going to move all of the @Assert constraints onto our input. Copy the two off of $title and move those into CheeseListingInput
We do need a use statement... but let's worry about that in a minute.
Copy the constraint from $description, move it
Copy the one from  delete it and... also delete the constraint from $description. We could also choose to keep these validation rules in our entity... which would make sense if we used this class outside of our API and it needed to be validated there.
Paste the constraint above price and... there's one more constraint above owner: @IsValidOwner()
Copy it, delete it, and move it into the input.
That's it! To get the use statements, re-type the end of NotBlank and hit tab to auto-complete it - that added the use statement on top - and do the same for IsValidOwner
Ok, cool! All of the validation rules live here and we have no constraints in CheeseListing.