45.2.UUID's
The UUID Doctrine Type
What this allows us to do - back in User - is add a new property - private $uuid - and, above this, say @ORM\Column() with type="uuid". That would not have worked before we installed that library and got the new config file. Also set this to unique=true
UUID's are strings, but the uuid type will make sure to store the UUID in the best possible way for whatever database system you're using. And when we query, it will turn that string back into a UUID object, which is ultimately what's stored on this property. You'll see that in a minute.
Could we know remove the auto-increment column and make this the primary key in Doctrine?
Yes, you could. But I won't. Why? Some databases - like MySQL - have performance problems with foreign keys when your primary key is a string. PostgreSQL does not have this problem, so do whatever is best for you. But there's no real disadvantage to having the auto-increment primary key, but a UUID as your API identifier.

Generating a Safe Migration
Ok: we added a new property. So let's generate the migration for it. Find your terminal and run:

  symfony console make:migration
Let's go check that out. Go into the migrations/ directory... and open the latest file
Since I'm using MySQL, you can see that it's storing this as a char field with a length of 36

The only tricky thing is that because we do already have a database with users in it - the fact that this column is NOT NULL will make the migration fail because the existing records will have no value.