45.3.UUID's
The only tricky thing is that because we do already have a database with users in it - the fact that this column is NOT NULL will make the migration fail because the existing records will have no value.
To fix this, temporarily change it to DEFAULT NULL.
Then, right after this, say $this->addSql() with UPDATE cheese_listing SET uuid = UUID()
That's a MySQL function to generate UUID's.