45.5.UUID's
Auto-Setting the UUID
Got it! So at this point, we have a new column... but nobody is using it. Let's run the user tests:

symfony php bin/phpunit tests/Functional/CheeseListingResourceTest.php
And... ah! It explodes like crazy! Of course:

  Column uuid cannot be null.

It's required in the database... but we're never setting it.

Unlike an auto-increment ID, the UUID is not automatically set, which is fine. We can set it ourselves in the constructor. Scroll down... we already have a constructor. Add $this->uuid = Uuid - auto-complete the one from Ramsey\ - then call uuid4(), which is how you get a random UUID string
Run the tests again:

  symfony php bin/phpunit tests/Functional/CheeseListingResourceTest.php
Now they're happy!
Next, the UUID is not part of our API at all yet. Let's tell API Platform to start using it as the identifier.