46.2.UUID as a API Identifier
Using the UUID as the API Identifier
So how do we tell API Platform to use the uuid property as the \identifier"? It's actually pretty simple! And we talked about it before.
Go to the top of User. Every API Resource needs an identifier. And when you use Doctrine, API Platform assumes that you want the database id as the identifier. To tell it to use something different, add @ApiProperty() with identifier=false
That says:

  Hey! Please don't use this as the identifier.

Then, above uuid, add @ApiProperty() with identifier=true
That's it! Try the test again:

  symfony php bin/phpunit --filter=testCreateUser
And... got it! But if we run all of the tests:

  symfony php bin/phpunit
Ah... things are not so good. It turns out that we were relying on the id as the identifier in a lot of places. Let's see an example. In testCreateCheeseListing(), we send the owner field set to /api/users/1. But... that is not the IRI of that user anymore! The IRI of every user just changed!